Create Old snapshots and low space volumes
This script connects to a controller and gets all snapshots older than the days specified as well as volumes greater than the specified percentage of usage. 

It can then either output this information into a csv file (one for snapshots and one for volumes) or can go through each snapshot in the list interactively and delete them one by one.